Dev mile stone 1

+++ Src/Trials/ConsoleApp/BL/Objects/CIFTYPE.cs
@@ -2,7 +2,9 @@ using System.Linq;
 using System;
 using System.Collections.Generic;
 using Microsoft.EntityFrameworkCore;
-using SIS.OpenCore.EL;
+using SIS.OpenCore.Model;
+using SIS.OpenCore.DAL;
+using SIS.OpenCore.DAL.Context;
 using SIS.OpenCore.BL;
 
 
@@ -10,7 +12,7 @@ namespace SIS.OpenCore.BL.Objects
 {
     public partial class CifTYPE
     {
-        public static bool ValidateExists(byte TypeCode)
+        public static bool ValidateExists(short TypeCode)
         {
             OpenCoreContext db = new OpenCoreContext();
             var Ret =   (from   t in db.LUT_CIF_TYPE
